Create non experts
Five year trend of users on SO. Selected only users created in March of 2013 and who last accessed SO this year, 2019 and had less than the average reputation score for the period (and more than for another set). This was for question posts. For answer posts, change posttypeid in line 6 to 2. Also change year in line 7 for the range of interest.